#41 update package cmd on main function for initialize http
